+++ packages/create-mechanic/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.6.2 - 2021-09-09
+
+### Fixed
+
+- Tweaked CLI language. Installation message shortened and tweaked "first" design function creation talk.
+
+## 0.6.1 - 2021-09-08
+
+### Fixed
+
+- Fixed dir name for p5-video template
+
 ## 0.6.0 - 2021-09-08
 
 ### Changed